Fix dbl render error on revoked token
If the refresh token route was hit, and the token was already revoked,
the code would render a JSON error, but it would not return. This
meant that the gem would continue executing and eventually
hit another render condition, causing a rails double render error.

Adding a return to this case fixes the issue.